filter 컴포넌트 관심사 분리(함수 및 상태 관리 로직 & Filter UI)
