+++ client/src/constants/initalBankData.ts
@@ -1,0 +1,14 @@
+import type { BankList } from "../type";
+
+export const initialBankData: BankList[] = [
+  { name: "부산은행", companyCode: "BS", checked: false },
+  { name: "씨티은행", companyCode: "CT", checked: false },
+  { name: "하나은행", companyCode: "HN", checked: false },
+  { name: "국민은행", companyCode: "KB", checked: false },
+  { name: "케이뱅크", companyCode: "KBK", checked: false },
+  { name: "카카오뱅크", companyCode: "KK", checked: false },
+  { name: "SC제일은행", companyCode: "SC", checked: false },
+  { name: "신한은행", companyCode: "SH", checked: false },
+  { name: "토스뱅크", companyCode: "TS", checked: false },
+  { name: "우리은행", companyCode: "WR", checked: false },
+];
