+++ client/src/constants/initialFilterState.ts
@@ -1,0 +1,14 @@
+export const initialFilterState = {
+  bank: {
+    isActive: false,
+    text: "은행명",
+  },
+  rate: {
+    isActive: false,
+    text: "기본금리순",
+  },
+  deposit: {
+    isActive: false,
+    text: "금액",
+  },
+};
