+++ client/src/hook/useFilterModalController.ts
@@ -1,0 +1,1 @@
+export const useFilterModalController = () => {};
