+++ client/src/utils/updateBankCheckState.ts
@@ -1,0 +1,19 @@
+import type { BankList } from "../type";
+
+export const updateBankCheckState = (
+  companyCode: string,
+  bankCheckState: BankList[]
+) => {
+  const updateBankList = bankCheckState.map((bank) => {
+    if (companyCode === bank.companyCode) {
+      return {
+        ...bank,
+        checked: !bank.checked,
+      };
+    } else {
+      return bank;
+    }
+  });
+
+  return updateBankList;
+};
